fix: 상세페이지에서의 이벤트 모달 위치 변경
